Move likelihood stuff to "simlike" module
